Crafter 3.0 Work
Updated Editorial Blueprint.
Changed the format for environment-config.xml per new design and moved it up a level. Left the old in there temporarily.
Changed SITENAME constant in all blueprints to be {siteName} for consistency and changed the code to handle the new name.
Changed default parameters as we continue to integrate with Search and Preview Deployer for more out-of-the-box functionality.