[5604] Rollback invoking EnvOverrides.getSiteValues
Roll back https://github.com/craftercms/studio/pull/2991 since getSiteValues is a little too aggressive in all that it does particularly for the entry page. Added a new method that does only what we needed. Should discuss if there's a better place to locate the new method or a better name for it.

craftercms/craftercms#5604